+++ src/extension/utils/getProgressUrl.ts
@@ -3,18 +3,19 @@
  *-------------------------------------------------------------------------------------------- */
 
 export function getProgressUrl(text: string): string | null {
-  if (!text.includes('You can follow up on the status:')) {
-    return null
-  }
+  // if (!text.includes('You can follow up on the status:')) {
+  //   return null
+  // }
 
   const urlRegex = /(https?:\/\/[^\s]+)/g
   const urlMatches = text.match(urlRegex)
 
+  console.log(urlMatches, 'urlMatch')
+
   if (urlMatches) {
-    const [url] = urlMatches
-    return url.includes('jobStatus')
-      ? url.replace('jobStatus', 'progress')
-      : null
+    const url = urlMatches[1]
+    console.log(url, 'urlMatch2')
+    return url.includes('output') ? url.replace('output', 'progress') : null
   }
   return null
 }
